refactor : 인덱싱 편리성을 위한 타입 변경

+++ src/main/java/com/example/team1_be/domain/Schedule/DTO/FixSchedule.java
@@ -9,6 +9,6 @@ public class FixSchedule {
     @AllArgsConstructor
     @NoArgsConstructor
     public static class Request {
-        private Long selection;
+        private int selection;
     }
 }
